Driver function pointers changed to const. Created HUB descriptors' structures and macros. Basic HUB driver state machine, up to reading the USB descriptor. Commented HUB methods in USBD for testing basic driver, this will be changed next.

+++ modules/usb/inc/usbd.h
@@ -171,7 +171,7 @@ int usb_device_find(
 int usb_stack_handle_hubs( usb_stack_t* pstack );
 #endif
 
-int usb_stack_new_addr( usb_stack_t* pstack );
+int16_t usb_stack_new_addr( usb_stack_t* pstack );
 
 int usb_stack_update_devices( usb_stack_t* pstack );
